refactor submissions
