+++ src/screens/speaker/components/talkCard/index.js
