+++ src/screens/speaker/components/talkCard/talkCard.container.js
