+++ src/screens/speaker/components/talkCard/talkCard.css
