+++ src/screens/speaker/components/talkCard/talkCard.jsx
