+++ src/screens/speaker/components/talksTable/index.js
