+++ src/screens/speaker/components/talksTable/talksTable.css
