+++ src/screens/speaker/submission/talkSubmission/index.js
@@ -1,0 +1,1 @@
+export { default } from './talkSubmission.container'
