+++ src/screens/speaker/submission/talksSelection/index.js
@@ -1,0 +1,1 @@
+export { default } from './talksSelection.container'
