+++ src/screens/speaker/submission/talksSelection/talksSelection.container.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import forRoute from 'hoc-little-router'
 
 import { getSpeakerAppEvent } from 'redux/ui/speaker'
-import TalksSubmission from './talksSubmission'
+import TalksSelection from './talksSelection'
 
 const mapState = (state) => {
   const { name } = getSpeakerAppEvent(state)
@@ -13,4 +13,4 @@ const mapState = (state) => {
 export default compose(
   forRoute('TALKS_SUBMISSION', { absolute: true }), //
   connect(mapState), //
-)(TalksSubmission)
+)(TalksSelection)
