+++ src/screens/speaker/submission/talksSelection/talksSelection.css
@@ -1,4 +1,4 @@
-.talks-submission .event-subtitle {
+.talks-selection .event-subtitle {
   color: var(--secondary-color);
   font-weight: 600;
 }
