+++ src/screens/speaker/submission/talksSelection/talksSelectionTable.container.js
@@ -6,7 +6,7 @@ import loader from 'hoc-react-loader/build/core'
 import { getSpeakerAppEvent } from 'redux/ui/speaker'
 import speakerTalks from 'redux/ui/speaker/talks'
 import LoadingIndicator from 'components/loading'
-import TalksTable from '../components/talksTable'
+import TalksTable from '../../components/talksTable'
 
 const mapState = (state) => {
   const { name } = getSpeakerAppEvent(state) || {}
