--- src/screens/speaker/talk/submit/index.js
@@ -1,1 +1,0 @@
-export { default } from './submit.container'
