--- src/screens/speaker/talks/talksSubmission/index.js
@@ -1,1 +1,0 @@
-export { default } from './talksSubmission.container'
